Reorganized to reflect config moving out of repo Added pull sh and bat files to pull latest code before running
