+++ src/DeviceModules/Pure.py
@@ -1,7 +1,7 @@
 import requests
 import json
 from loguru import logger
-from config import classes
+from DeviceModules import classes
 
 GB = ((1/1024)/1024)
