+++ src/pull.bat
@@ -1,0 +1,4 @@
+@echo off
+git fetch origin
+git pull
+python wrapper.py
