+++ src/pull.sh
@@ -1,0 +1,4 @@
+#!/bin/bash -e
+git fetch origin
+git pull
+python test.py
